백엔드 마무리

+++ server/articles/views.py
@@ -6,12 +6,13 @@ from rest_framework.decorators import api_view
 
 # permission Decorators
 from rest_framework.decorators import permission_classes
-from rest_framework.permissions import IsAuthenticatedOrReadOnly
+from rest_framework.permissions import IsAuthenticatedOrReadOnly, IsAuthenticated
 
 from rest_framework import status
 from django.shortcuts import get_object_or_404, get_list_or_404
 from .serializers import ArticleListSerializer, ArticleSerializer, CommentSerializer, ArticleDetailSerializer, ArticleUpdateSerializer
 from .models import Article, Comment
+# from accounts.models import User
 
 
 
@@ -22,7 +23,7 @@ def article_list(request):
         articles = Article.objects.all()
         # articles = get_list_or_404(Article)
         serializer = ArticleListSerializer(articles, many=True)
-        return Response(serializer.data)
+        return Response(serializer.data, status=status.HTTP_200_OK)
 
     elif request.method == 'POST':
         serializer = ArticleSerializer(data=request.data)
@@ -32,41 +33,53 @@ def article_list(request):
 
 
 @api_view(['GET', 'DELETE', 'PUT'])
+@permission_classes([IsAuthenticatedOrReadOnly])
 def article_detail(request, article_pk):
     article = Article.objects.get(pk=article_pk)
     # article = get_object_or_404(Article, pk=article_pk)
 
     if request.method == 'GET':
         serializer = ArticleDetailSerializer(article)
-        return Response(serializer.data)
+        return Response(serializer.data, status=status.HTTP_200_OK)
     
     elif request.method == 'DELETE':
-        article.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if request.user == article.user:
+            article.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(status=status.HTTP_403_FORBIDDEN)
 
     elif request.method == 'PUT':
-        serializer = ArticleUpdateSerializer(article, data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data)
+        if request.user == article.user:
+            serializer = ArticleUpdateSerializer(instance=article, data=request.data)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+                return Response(serializer.data)
+        return Response(status=status.HTTP_403_FORBIDDEN)
 
 
 @api_view(['GET','POST'])
+@permission_classes([IsAuthenticatedOrReadOnly])
 def comment_list(request, article_pk):
     if request.method == 'GET':
         comments = Comment.objects.filter(article=article_pk)
         # comments = get_list_or_404(Comment)
         serializer = CommentSerializer(comments, many=True)
-        return Response(serializer.data)
+        return Response(serializer.data, status=status.HTTP_200_OK)
+    
     elif request.method == 'POST':
+        article = Article.objects.get(pk=article_pk)
         serializer = CommentSerializer(data=request.data)
         if serializer.is_valid(raise_exception=True):
-            serializer.save(user=request.user, article=article_pk)
+            serializer.save(user=request.user, article=article)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
 
+
 @api_view(['DELETE', 'PUT'])
+@permission_classes([IsAuthenticated])
 def comment_update(request, comment_pk):
     comment = Comment.objects.get(pk=comment_pk)
+    if comment.user != request.user:
+        return Response(status=status.HTTP_403_FORBIDDEN)
     # comment = get_object_or_404(Comment, pk=comment_pk)
     if request.method == 'DELETE':
         comment.delete()
@@ -77,3 +90,15 @@ def comment_update(request, comment_pk):
         if serializer.is_valid(raise_exception=True):
             serializer.save()
             return Response(serializer.data)
+        
+
+@api_view(['PUT'])
+@permission_classes([IsAuthenticated])
+def likes(request, article_pk):
+    article = Article.objects.get(pk=article_pk)
+
+    if article.like_users.filter(pk=request.user.pk).exists():
+        article.like_users.remove(request.user)
+    else:
+        article.like_users.add(request.user)
+    return Response(status=status.HTTP_200_OK)
